refactor: Move displayCurrentProblem to reusable helpers
