+++ src/workflowUtils.js
@@ -1,4 +1,25 @@
 const { getNamedRangeValue } = require("./sheetUtils/getNamedRangeValue");
+const { getInputsFromSheetUI } = require("./sheetUtils/getInputsFromSheetUI");
+const { setInputsOnSheetUI } = require("./sheetUtils/setInputsOnSheetUI");
+
+/**
+ * Updates the UI with the selected problem's attributes.
+ *
+ * Retrieves the input fields map from the UI, maps problem attributes to the corresponding keys,  
+ * and writes the updated values back to the UI.
+ *
+ * @param {Object} problemAttributes - An object containing attribute key-value pairs for the selected problem.
+ */
+function displayCurrentProblem(problemAttributes) {
+    const inputsMap = getInputsFromSheetUI('ControlPanel_CurrentProblemInputs');
+    for (const key of inputsMap.keys()) {
+        if (problemAttributes.hasOwnProperty(key)) {
+            inputsMap.set(key, problemAttributes[key]);
+        }
+    }
+
+    setInputsOnSheetUI('ControlPanel_CurrentProblemInputs', inputsMap);
+}
 
 /**
  * Retrieves the current problem's LeetCode ID from the named range.
@@ -39,4 +60,4 @@ function isAttemptDone() {
     return getNamedRangeValue('ControlPanel_EndTime') != '';
 }
 
-module.exports = { getCurrentProblemLcId, isAttemptInProgress, isAttemptDone }
+module.exports = { displayCurrentProblem, getCurrentProblemLcId, isAttemptInProgress, isAttemptDone }
